Added full support to postgres

+++ api/src/main/java/es/urjc/videotranscoding/repository/OriginalRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import es.urjc.videotranscoding.entities.Original;
 
 @Repository
-public interface OriginalRepository extends PagingAndSortingRepository<Original, Long> {
+public interface OriginalRepository extends PagingAndSortingRepository<Original, Integer> {
 	Page<Original> findAllByUser(Pageable page, String user);
 
 	Page<Original> findAll(Pageable pageable);
